azure_storage_azure_function/requirements.txt
azure-functions
azure-storage-blob
azure-eventhub
